+++ src/frontend/SourceManager.cpp
@@ -108,6 +108,7 @@ namespace panther{
 	};
 
 
+	// TODO: optimize base types
 	auto SourceManager::initIntrinsics() noexcept -> void {
 		evo::debugAssert(this->isLocked(), "Can only initialize intrinsics when locked");
 		
@@ -116,13 +117,18 @@ namespace panther{
 			base_type.call_operators.emplace_back(std::vector<PIR::Type::ID>{}, std::nullopt);
 
 			const PIR::BaseType::ID base_type_id = this->createBaseType(std::move(base_type));
-
-			// this->intrinsics.tokens.emplace_back(Token::Kind::Intrinsic, "__printHelloWorld");
-
-			// this->intrinsics.funcs.emplace_back(Token::ID( uint32_t(this->intrinsics.tokens.size() - 1) ), base_type_id, std::nullopt, false);
-
 			this->intrinsics.emplace_back(PIR::Intrinsic::Kind::__printHelloWorld, "__printHelloWorld", base_type_id);
 		}
+
+
+		{ // breakpoint()
+			auto base_type = PIR::BaseType(PIR::BaseType::Kind::Function);
+			base_type.call_operators.emplace_back(std::vector<PIR::Type::ID>{}, std::nullopt);
+
+			const PIR::BaseType::ID base_type_id = this->createBaseType(std::move(base_type));
+			this->intrinsics.emplace_back(PIR::Intrinsic::Kind::breakpoint, "breakpoint", base_type_id);
+		}
+
 	};